Moved business logic function to specific files
Now the output, logic, api, data, input and settings functions
are in separated files, allowing us to manage the project better
