+++ greenpass/URLs.py
@@ -1,0 +1,4 @@
+# Base url to retrieve data (DGC)
+BASE_URL_DGC   = "https://get.dgc.gov.it/v1/dgc/"
+# Base url to retrieve data (NHS)
+BASE_URL_NHS   = "https://covid-status.service.nhsx.nhs.uk/"
